Refactor: Flatten parser and assembly modules
Flatten `src/parser/` directory into `src/parser.rs`, `src/grammar.rs`, and `src/grammar.pest`.
Flatten `src/semantic/assembly/` into `src/semantic/assembly.rs`.
Delete empty directories.
Update `src/lib.rs` and imports accordingly.
Update `.jules/razor.md` with reduction log.
